requirements.txt
numpy
opencv-python
matplotlib
scikit-image
lpips
torchmetrics
ptwt
PyWavelets